# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "mousecake-go 区块链后端服务 API 文档，包含用户认证、报价聚合等模块。",
        "title": "mousecake-go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/providers": {
            "get": {
                "description": "获取当前可用的报价供应商列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "报价"
                ],
                "summary": "获取供应商列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
        "/quote": {
            "get": {
                "description": "根据指定参数从供应商获取代币兑换报价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "报价"
                ],
                "summary": "获取报价",
                "parameters": [
                    {
                        "enum": [
                            "okx",
                            "zerox"
                        ],
                        "type": "string",
                        "description": "供应商名称",
                        "name": "provider",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "链 ID",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE",
                        "description": "源代币地址",
                        "name": "from_token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
                        "description": "目标代币地址",
                        "name": "to_token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "1000000000000000000",
                        "description": "兑换数量",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "exactIn",
                            "exactOut"
                        ],
                        "type": "string",
                        "description": "兑换模式",
                        "name": "swap_mode",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
        "/swap": {
            "post": {
                "description": "根据报价参数获取链上交易所需的交易数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "报价"
                ],
                "summary": "获取交易数据",
                "parameters": [
                    {
                        "description": "获取交易数据请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_quote.getSwapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
        "/swap/{id}/submit": {
            "post": {
                "description": "提交已广播的链上交易哈希，完成 swap 记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "报价"
                ],
                "summary": "提交交易哈希",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Swap 记录 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "交易哈希",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_quote.submitSwapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 业务状态码，0 表示成功，非 0 表示失败。",
                    "type": "integer"
                },
                "data": {
                    "description": "Data 响应数据载荷，成功时为业务数据，失败时为 nil。"
                },
                "message": {
                    "description": "Message 响应消息描述。",
                    "type": "string"
                }
            }
        },
...
        "internal_quote.getSwapRequest": {
            "type": "object",
            "required": [
                "amount",
                "chain_id",
                "from_token",
                "provider",
                "swap_mode",
                "to_token",
                "user_wallet_address"
            ],
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "1000000000000000000"
                },
                "chain_id": {
                    "type": "integer",
                    "example": 1
                },
                "from_token": {
                    "type": "string",
                    "example": "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE"
                },
                "provider": {
                    "type": "string",
                    "example": "okx"
                },
                "slippage_percent": {
                    "type": "number",
                    "example": 0.5
                },
                "swap_mode": {
                    "type": "string",
                    "example": "exactIn"
                },
                "to_token": {
                    "type": "string",
                    "example": "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48"
                },
                "user_wallet_address": {
                    "type": "string",
                    "example": "0x1234567890abcdef1234567890abcdef12345678"
                }
            }
        },
        "internal_quote.submitSwapRequest": {
            "type": "object",
            "required": [
                "tx_hash"
            ],
            "properties": {
                "tx_hash": {
                    "type": "string",
                    "example": "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}